requirements.txt
Flask==2.2.5
Flask_JWT_Extended==4.5.2
Flask_Migrate==4.0.5
Flask_SQLAlchemy==2.5.1
pandas==1.5.2
PyJWT==2.8.0
Requests==2.31.0
SQLAlchemy==1.4.49
Werkzeug==2.2.3
